[TE] alerter - refactor internal base class for reuable utilities
